Fix race condition in blocking()
Currently there is a race condition in `blocking()`.

The code is waiting for the `m.signal` condition to fire in a goroutine,
and that routine is supposed to be stopped in case of timeouts or context
termination with an extra `Broadcast()` call.

When the global server context is already done upon entering that
function however, the broadcast will race against the scheduling of the
goroutine, and in case it misses its target, the waitgroup operation will
never finish and the process deadlocks.

To fix this, turn the logic on its head and wait for the condition
synchronously, and send a signal from a goroutine in case we hit a timeout
or the context is terminated.

This isn't an artificial case. The bug can be reproduced with go-redis/v8
and the following snippet:

```
package main

import (
	"context"
	"fmt"
	"sync"

	"github.com/alicebob/miniredis/v2"
	"github.com/go-redis/redis/v8"
)

func main() {
	mini, err := miniredis.Run()
	if err != nil {
		panic(err)
	}

	redis := redis.NewUniversalClient(&redis.UniversalOptions{
		Addrs: []string{mini.Addr()},
	})

	ctx, cancel := context.WithCancel(redis.Context())

	redis.LPush(ctx, "x", "a")

	var wg sync.WaitGroup
	wg.Add(1)

	go func() {
		for {
			_, err := redis.BRPop(ctx, 0, "x").Result()
			if err == nil {
				wg.Done()
			}
		}
	}()

	fmt.Printf("Waiting ...\n")
	wg.Wait()

	fmt.Printf("Cancelling ...\n")
	cancel()
	redis.Close()
	mini.Close()

	fmt.Printf("Done.\n")
}

```