Retry broadcast until the receiver is unblocked
The last patch made the race condition between the signal `Wait()` and the
broadcaster that fires when a timeout happens or when the context is going
down more unlikely. It is however theoretically still possible to run into
that.

To fix this properly, repeat the broadcast until the receiver has atomically
flipped a switch after it received the signal.