Fix possibility to generate duplicate IDs in XADD
The existing code is attempting to avoid duplicate IDs by checking the
last ID in the queue. But if the queue is emptied and a new event
added fast enough, it's possible to end up with duplicate IDs.

This adds a variable to streamKey that keeps track of the last ID even
when the stream itself is empty, which likely solves most cases where
this could cause problems.