Move .PHONY next to each target
It's clearer to keep target and it's .PHONY together. When reading
Makefile it's easier to see if target is phony or not (without jumping
back to first line and reading all of .PHONY dependencies). It's also
harder to forget remove/rename .PHONY dependency when target is changed
when they are together.

There is no official standard but make documentation[1] in .DEFAULT_GOAL
use multiple .PHONY:
> .PHONY: foo
> foo: ; @echo $@
>
> $(warning default goal is $(.DEFAULT_GOAL))
>
> # Reset the default goal.
> .DEFAULT_GOAL :=
>
> .PHONY: bar
> bar: ; @echo $@

PR: https://github.com/alicebob/miniredis/issues/371
[1]: https://www.gnu.org/software/make/manual/html_node/Special-Variables.html

Signed-off-by: Wojciech Szarański <wojciech.szaranski@gmail.com>